dev-requirements.txt
beautifulsoup4>=4.9.3
pytest-ckan
